feat: make Config + Pauser + RewardManager + Staking TransparentUpgProxies

+++ test/utils/Defaults.sol
@@ -8,7 +8,7 @@ contract Defaults {
     uint256 public constant EPOCH_DURATION = 1 days;
     uint256 public constant FIRST_EPOCH = 1;
     uint256 public constant PERFORMANCE_SCALE = 1e18;
-    uint256 public constant MAX_NODES = 50;
+    uint8 public constant MAX_NODES = 50;
     uint256 public constant MIN_STAKE = 100_000 ether;
     uint8 public constant MAX_REGISTRY_NODES = 50;
 }
